pyproject.toml
[project]
name = "ds-storm-impact-harmonisation"
version = "0.1.0"
description = "Harmonising storm impact data"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "exactextract>=0.3.0",
    "great-tables>=0.21.0",
    "marimo>=0.9",
    "matplotlib>=3.10.8",
    "ocha-relay",
    "ocha-stratus>=0.1.7",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "plotnine>=0.15.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32",
    "scipy>=1.17.1",
    "statsmodels>=0.14",
]

[dependency-groups]
dev = [
    "contextily>=1.7.0",
    "jupyter>=1.1.1",
    "plotly>=6.7.0",
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.isort]
known-first-party = ["ds_storm_impact_harmonisation"]

[tool.uv.sources]
ocha-relay = { git = "https://github.com/OCHA-DAP/ocha-relay.git", rev = "v0.2.0" }
